fix: address KDL edge cases and modularize codebase
