+++ extensions/niri-monitors/src/global-utils.ts
